requirements.txt
numpy
pandas
matplotlib
seaborn
tensorflow.keras
keras
